pyproject.toml
[build-system]
requires = ["setuptools >= 75.3", "setuptools_scm[toml]"]
build-backend = "setuptools.build_meta"

[project]
name = "emerald-directory"
dynamic = ['version']
description = "Community Support Directory/Forum"
urls = { "homepage" = "https://github.com/induane" }
authors = [
    { "name" = "Brant Watson", "email" = "oldspiceap@gmail.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.13"
dependencies = [
    "Django~=6.0",
    "boltons~=25.0",
    # "diskcache~=5.6",
    # "docutils~=0.20",
    "pillow>=12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    # "django-extensions",
    # "mypy",
    # "log-color==2.0.0",
    "types-boltons~=25.0",
    # "types-docutils~=0.20",
]
#docs = ["sphinx", "sphinx_rtd_theme", "pygraphviz", "pyparsing", "jinja2<3.1"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"emerald_heart" = ["py.typed"]
"*" = [
    "*.css",
    "*.eot",
    "*.html",
    "*.ico",
    "*.ini",
    "*.jpg",
    "*.js",
    "*.json",
    "*.png",
    "*.png",
    "*.svg",
    "*.ttf",
    "*.webp",
    "*.woff",
    "*.woff2",
]


[tool.setuptools_scm]
root = "."
write_to = "src/emerald_heart/_version.py"

[tool.ruff]
line-length = 120
exclude = ["*.pyi", "_version.py"]

[tool.ruff.lint]
select = [
    "E", # Pycodestyle Errors
    "W", # Pycodestyle Warnings
    "F", # PyFlakes rules
    "B", # Bugbear rules (likely bugs & design problems in your code)
    "N", # Pep8 naming rules
    "D", # Pydocstyle (rules for docstrings)
    "I", # Import sorting
]
ignore = [
    "D102", # Some functions are pretty self documenting
    "D105", # Documenting every magic method is tedious and often redundant/self-explanatory
    "D203", # D203 & D211 are incompatible; have to choose one or the other
    "D212", # D212 and D213 are incompatible; D212 makes visual alignment weird
    "N818", # Don't require exception subclass names to all end with 'Error'
    "D104", # Don't require a docstring for every modulee __init__ file
    "D107", # Don't require a docstring for every __init__ method
    "D100", # Don't make all modules include a docstring
]

# Allow unused variables named "_".
dummy-variable-rgx = "^_$"

# Select the classes of fixable errors
fixable = ["E", "F", "W", "N", "D", "I"]

unfixable = [
    "B",     # Avoid trying to fix flake8-bugbear (`B`) violations.
    "UP013", # Don't rewrite TypedDict functions as classes
]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = [
    "D101", # Relax some docstrings rules
    "D102",
    "N802", # Relax function name rules
    "N806", # Relax class naming rules
    "E501", # Long lines are par for the couse in Unitests
    "N805", # Allow names other than "self" for instances methods
    "B009", # Relax getattr rules
    "B018", # Allow useless expressions (often used to assert calls raise exceptions)
    "I002", # Don't require tests to use required imports
]
"**/settings/*.py" = ["F405"]
"**/hints.py" = ["D101"]
"**/migrations/*.py" = ["E501", "D101"]

[[tool.ty.overrides]]
exclude = ["**/utils/image.py"]

[tool.ty.overrides.rules]
possibly-unresolved-reference = "warn"
possibly-missing-attribute = "ignore"
